refactor: productContext is now inside productList directory
