+++ src/app/product/[id].tsx
@@ -1,7 +1,7 @@
 import AddToCartButton from "@/components/AddToCartButton";
 import ParallaxScrollView from "@/components/ParallaxScrollView";
 import { useCartContext } from "@/contexts/cartContext/CartContext";
-import { useProductContext } from "@/contexts/ProductContext";
+import { useProductContext } from "@/contexts/productList/ProductContext";
 import { useLocalSearchParams, useRouter } from "expo-router";
 import React from "react";
 import { Image, StyleSheet, Text } from "react-native";
